requirements-dev.txt
python-telegram-bot==13.8.1
Django==4.0.2
psycopg2-binary>=2.8
matplotlib==3.5.1